correct named object for test, xid with _TEST, abbreviation of names;…

+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/page/datasource/CreateDataSourcePageTest.java
@@ -3,7 +3,6 @@ package org.scadalts.e2e.test.impl.tests.page.datasource;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.DataSourceCriteria;
 import org.scadalts.e2e.page.impl.criterias.IdentifierObjectFactory;
 import org.scadalts.e2e.page.impl.criterias.identifiers.DataSourceIdentifier;
@@ -11,13 +10,11 @@ import org.scadalts.e2e.page.impl.dicts.DataSourceType;
 import org.scadalts.e2e.page.impl.dicts.UpdatePeriodType;
 import org.scadalts.e2e.page.impl.pages.datasource.DataSourcesPage;
 import org.scadalts.e2e.test.impl.creators.DataSourcePointObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.StringContains.containsString;
 
-@RunWith(TestWithPageRunner.class)
 public class CreateDataSourcePageTest {
 
     private static final DataSourceType dataSourceType = DataSourceType.VIRTUAL_DATA_SOURCE;
@@ -30,13 +27,14 @@ public class CreateDataSourcePageTest {
     @Before
     public void setup() {
         DataSourceCriteria criteria = DataSourceCriteria.criteria(dataSourceIdentifier,updatePeriodType);
-        dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(TestWithPageUtil.getNavigationPage(), criteria);
+        dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(TestWithPageUtil.openNavigationPage(), criteria);
         dataSourcesPageSubject = dataSourcePointObjectsCreator.openPage();
     }
 
     @After
     public void clean() {
-        dataSourcePointObjectsCreator.deleteObjects();
+        if(dataSourcePointObjectsCreator != null)
+            dataSourcePointObjectsCreator.deleteObjects();
     }
 
     @Test
