+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/page/datasource/DeleteDataSourcePageTest.java
@@ -3,7 +3,6 @@ package org.scadalts.e2e.test.impl.tests.page.datasource;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.DataSourceCriteria;
 import org.scadalts.e2e.page.impl.criterias.IdentifierObjectFactory;
 import org.scadalts.e2e.page.impl.criterias.identifiers.DataSourceIdentifier;
@@ -11,14 +10,12 @@ import org.scadalts.e2e.page.impl.dicts.DataSourceType;
 import org.scadalts.e2e.page.impl.dicts.UpdatePeriodType;
 import org.scadalts.e2e.page.impl.pages.datasource.DataSourcesPage;
 import org.scadalts.e2e.test.impl.creators.DataSourcePointObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.StringContains.containsString;
 
-@RunWith(TestWithPageRunner.class)
 public class DeleteDataSourcePageTest {
 
     private DataSourceIdentifier dataSourceToDeleteName = IdentifierObjectFactory
@@ -33,14 +30,15 @@ public class DeleteDataSourcePageTest {
 
         dataSourceToDeleteCriteria = DataSourceCriteria.criteria(dataSourceToDeleteName,
                 UpdatePeriodType.SECOND);
-        dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(TestWithPageUtil.getNavigationPage(), dataSourceToDeleteCriteria);
+        dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(TestWithPageUtil.openNavigationPage(), dataSourceToDeleteCriteria);
         dataSourcePointObjectsCreator.createDataSources();
         dataSourcesPageSubject = dataSourcePointObjectsCreator.openPage();
     }
 
     @After
     public void clean() {
-        dataSourcePointObjectsCreator.deleteObjects();
+        if(dataSourcePointObjectsCreator != null)
+            dataSourcePointObjectsCreator.deleteObjects();
     }
 
     @Test
