+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/page/datasource/EditDataSourcePageTest.java
@@ -3,19 +3,16 @@ package org.scadalts.e2e.test.impl.tests.page.datasource;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.DataSourceCriteria;
 import org.scadalts.e2e.page.impl.dicts.UpdatePeriodType;
 import org.scadalts.e2e.page.impl.pages.datasource.DataSourcesPage;
 import org.scadalts.e2e.page.impl.pages.datasource.EditDataSourceWithPointListPage;
 import org.scadalts.e2e.test.impl.creators.DataSourcePointObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 
-@RunWith(TestWithPageRunner.class)
 public class EditDataSourcePageTest {
 
     private DataSourceCriteria criteria;
@@ -27,14 +24,15 @@ public class EditDataSourcePageTest {
     @Before
     public void createDataSource() {
         criteria = DataSourceCriteria.virtualDataSource(UpdatePeriodType.SECOND, 13);
-        dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(TestWithPageUtil.getNavigationPage(), criteria);
+        dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(TestWithPageUtil.openNavigationPage(), criteria);
         dataSourcesPage = dataSourcePointObjectsCreator.openPage();
         editDataSourceWithPointListPageSubject = dataSourcePointObjectsCreator.createDataSources();
     }
 
     @After
     public void clean() {
-        dataSourcePointObjectsCreator.deleteObjects();
+        if(dataSourcePointObjectsCreator != null)
+            dataSourcePointObjectsCreator.deleteObjects();
     }
 
     @Test
@@ -68,7 +66,7 @@ public class EditDataSourcePageTest {
     public void test_edit_update_period_type_in_data_source() {
 
         //given:
-        UpdatePeriodType updatePeriodsTypeExp = UpdatePeriodType.MILLISECOUND;
+        UpdatePeriodType updatePeriodsTypeExp = UpdatePeriodType.MILLISECOND;
 
         //when
         UpdatePeriodType updatePeriodTypeBefore = editDataSourceWithPointListPageSubject
