+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/page/pointlinks/DeletePointLinksPageTest.java
@@ -8,14 +8,13 @@ import org.scadalts.e2e.page.impl.criterias.PointLinkCriteria;
 import org.scadalts.e2e.page.impl.dicts.EventType;
 import org.scadalts.e2e.page.impl.pages.pointlinks.PointLinksPage;
 import org.scadalts.e2e.test.impl.creators.AllObjectsForPointLinkTestCreator;
-import org.scadalts.e2e.test.impl.runners.TestParameterizedWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.not;
 
-@RunWith(TestParameterizedWithPageRunner.class)
+@RunWith(Parameterized.class)
 public class DeletePointLinksPageTest {
 
     @Parameterized.Parameters(name = "{index}: script: {0}, {1}")
@@ -41,7 +40,7 @@ public class DeletePointLinksPageTest {
 
     @BeforeClass
     public static void setup() {
-        allObjectsForPointLinkTestCreator = new AllObjectsForPointLinkTestCreator(TestWithPageUtil.getNavigationPage());
+        allObjectsForPointLinkTestCreator = new AllObjectsForPointLinkTestCreator(TestWithPageUtil.openNavigationPage());
         allObjectsForPointLinkTestCreator.getDataSourcePointObjectsCreator().createObjects();
         pointLinksPageSubject = allObjectsForPointLinkTestCreator.openPage();
         criteria = allObjectsForPointLinkTestCreator.getCriteria();
@@ -49,7 +48,8 @@ public class DeletePointLinksPageTest {
 
     @AfterClass
     public static void clean() {
-        allObjectsForPointLinkTestCreator.deleteObjects();
+        if(allObjectsForPointLinkTestCreator != null)
+            allObjectsForPointLinkTestCreator.deleteObjects();
     }
 
     @Before
@@ -64,7 +64,8 @@ public class DeletePointLinksPageTest {
 
     @After
     public void deletePointLink() {
-        allObjectsForPointLinkTestCreator.getPointLinksObjectsCreator().deleteObjects();
+        if(allObjectsForPointLinkTestCreator != null)
+            allObjectsForPointLinkTestCreator.getPointLinksObjectsCreator().deleteObjects();
     }
 
     @Test
