+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/service/storungs/live/GetAllLivesParametersServiceTest.java
@@ -4,7 +4,6 @@ import lombok.extern.log4j.Log4j2;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.DataPointCriteria;
 import org.scadalts.e2e.page.impl.criterias.IdentifierObjectFactory;
 import org.scadalts.e2e.page.impl.criterias.properties.DataPointLoggingProperties;
@@ -14,7 +13,6 @@ import org.scadalts.e2e.page.impl.pages.navigation.NavigationPage;
 import org.scadalts.e2e.service.impl.services.storungs.PaginationParams;
 import org.scadalts.e2e.service.impl.services.storungs.StorungAlarmResponse;
 import org.scadalts.e2e.test.impl.creators.StorungsAndAlarmsObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.RegexUtil;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
@@ -31,9 +29,9 @@ import static org.hamcrest.core.AnyOf.anyOf;
 import static org.junit.Assert.*;
 import static org.scadalts.e2e.test.impl.utils.StorungsAndAlarmsUtil.EXPECTED_DATE_ISO;
 import static org.scadalts.e2e.test.impl.utils.StorungsAndAlarmsUtil.getStorungsAndAlarms;
+import static org.scadalts.e2e.test.impl.utils.StorungsAndAlarmsUtil.sleep;
 
 @Log4j2
-@RunWith(TestWithPageRunner.class)
 public class GetAllLivesParametersServiceTest {
 
     private static PaginationParams paginationParams = PaginationParams.all();
@@ -61,7 +59,7 @@ public class GetAllLivesParametersServiceTest {
                 DataPointCriteria.noChange(IdentifierObjectFactory.dataPointNotifierBinaryTypeName(DataPointNotifierType.STORUNG), "1", DataPointLoggingProperties.tsChange()),
         };
 
-        NavigationPage navigationPage = TestWithPageUtil.getNavigationPage();
+        NavigationPage navigationPage = TestWithPageUtil.openNavigationPage();
         storungsAndAlarmsObjectsCreator = new StorungsAndAlarmsObjectsCreator(navigationPage, points);
         storungsAndAlarmsObjectsCreator.createObjects();
         storungsAndAlarmsObjectsCreator.setDataPointValue(1)
@@ -74,7 +72,8 @@ public class GetAllLivesParametersServiceTest {
 
     @AfterClass
     public static void clean() {
-        storungsAndAlarmsObjectsCreator.deleteObjects();
+        if(storungsAndAlarmsObjectsCreator != null)
+            storungsAndAlarmsObjectsCreator.deleteObjects();
     }
